requirements.txt
customtkinter>=5.0.0
mysql-connector-python>=8.0.33
Pillow>=9.0.0
tkcalendar>=1.6.1

